requirements.txt
Django==2.1.7
django-cors-headers==2.4.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
mysqlclient==1.4.2.post1
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2018.9
